+++ src/app/app.ts
@@ -7,17 +7,17 @@ import { AppController } from "./controllers/app.controller";
 import { ErrorMiddleware } from "./middleware/error.middleware";
 import { HandlebarsMiddleware } from "./middleware/handlebars.middleware";
 
-const {
-  initializeApp,
-  applicationDefault,
-  cert,
-} = require("firebase-admin/app");
-const {
-  getFirestore,
-  Timestamp,
-  FieldValue,
-  Filter,
-} = require("firebase-admin/firestore");
+// const {
+//   initializeApp,
+//   applicationDefault,
+//   cert,
+// } = require("firebase-admin/app");
+// const {
+//   getFirestore,
+//   Timestamp,
+//   FieldValue,
+//   Filter,
+// } = require("firebase-admin/firestore");
 
 class App {
   // Create an instance of express, called "app"
@@ -46,18 +46,18 @@ class App {
     this.app.use(this.appController.router);
     this.app.use(this.errorMiddleware.router);
 
-    const firebaseConfig = {
-      apiKey: "AIzaSyC2EVM2_4lOG83k3kf_2XkylX-hOLc7Pfc",
-      authDomain: "pspmmfdw-6586b.firebaseapp.com",
-      projectId: "pspmmfdw-6586b",
-      storageBucket: "pspmmfdw-6586b.appspot.com",
-      messagingSenderId: "949890250152",
-      appId: "1:949890250152:web:81bb4fac772a0cc547d99d",
-      measurementId: "G-CX84FKXHHK",
-    };
-    const app = initializeApp(firebaseConfig);
+    // const firebaseConfig = {
+    //   apiKey: "AIzaSyC2EVM2_4lOG83k3kf_2XkylX-hOLc7Pfc",
+    //   authDomain: "pspmmfdw-6586b.firebaseapp.com",
+    //   projectId: "pspmmfdw-6586b",
+    //   storageBucket: "pspmmfdw-6586b.appspot.com",
+    //   messagingSenderId: "949890250152",
+    //   appId: "1:949890250152:web:81bb4fac772a0cc547d99d",
+    //   measurementId: "G-CX84FKXHHK",
+    // };
+    // const app = initializeApp(firebaseConfig);
 
-    const db = getFirestore();
+    // const db = getFirestore();
   }
 
   public listen() {
