+++ src/app/controllers/passport.js
@@ -4,6 +4,7 @@ const GOOGLE_CLIENT_ID =
   "701590633897-s9qh400b2vmubcv0611vjl0ns3qn31p5.apps.googleusercontent.com";
 const GOOGLE_CLIENT_SECRET = "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";
 import * as App from "../app.ts";
+import db from "../firebase.ts";
 
 passport.serializeUser(function (user, done) {
   done(null, user);
@@ -21,7 +22,7 @@ passport.use(
       callbackURL: "http://localhost:3000/google/callback",
     },
     function (profile, cb) {
-      const userRef = App.db.collection("users").doc(profile.id);
+      const userRef = db.collection("users").doc(profile.id);
       userRef.get().then((doc) => {
         if (!doc.exists) {
           const newUser = {
